Choosing Your Mobile Browser
Created a list of mobile browser which are privacy respecting and quite secure. it includes Mulch, Vanadium, Brave and Tor/Mull with a note.

Guide to Divest Recommended Browser Settings is also linked if someone wants to harden Brave and Mull. A note is added for Gecko Browsers *Firefox-based Android browsers lack per-site process isolation*.

Also it includes answers to the following questions

- Why use Vanadium over Brave Browser?
- Why don’t we recommend DuckDuckGo (DDG) Browser?

**Safari is pending. This guide is completed for Android Browsers.** Although I have added a sentence at the start for using Safari with Apple Private Relay.

Signed-off-by: sky768 <153304006+sky768@users.noreply.github.com>